chore: Refactor the Mana Calculator
