+++ src/components/ManaCalculator/constants.ts
@@ -1,0 +1,12 @@
+export const IOTA_SUPPLY = 4600000000000000;
+export const SHIMMER_SUPPLY = 1813620509000000;
+export const SLOT_DURATION = 10;
+export const SLOTS_IN_EPOCH = 8192;
+export const EPOCH_DURATION = SLOTS_IN_EPOCH * SLOT_DURATION;
+export const SECONDS_IN_YEAR = 60 * 60 * 24 * 365;
+export const EPOCH_DURATION_IN_YEARS = EPOCH_DURATION / SECONDS_IN_YEAR;
+export const BETA_PER_YEAR = 1 / 3;
+export const GENERATION_PER_SLOT = Math.pow(2, -17);
+export const BOOTSTRAPPING_DURATION = 1154;
+export const REWARDS_MANA_SHARE_COEFFICIENT = 2;
+export const EPOCH = 1154 + 1;
