+++ src/components/ManaCalculator/enums/parameters.enum.ts
@@ -1,0 +1,15 @@
+export enum UserType {
+    DELEGATOR,
+    VALIDATOR,
+  }
+  
+  export enum NetworkType {
+    IOTA,
+    SHIMMER,
+  }
+
+  export enum CongestionType {
+    LOW = 100000,
+    MEDIUM = 9000000,
+    HIGH = 500000000,
+  }
