+++ src/components/ManaCalculator/utils.ts
@@ -1,39 +1,15 @@
-import { NetworkType } from "./types";
+import { BETA_PER_YEAR, BOOTSTRAPPING_DURATION, EPOCH_DURATION_IN_YEARS, GENERATION_PER_SLOT, IOTA_SUPPLY, REWARDS_MANA_SHARE_COEFFICIENT, SHIMMER_SUPPLY, SLOTS_IN_EPOCH } from "./constants";
+import { NetworkType } from "./enums";
 
-// Constants
-const IOTA_SUPPLY = 4600000000000000;
-const SHIMMER_SUPPLY = 1813620509000000;
-const SLOT_DURATION = 10;
-const SLOTS_IN_EPOCH = 8192;
-export const EPOCH_DURATION = SLOTS_IN_EPOCH * SLOT_DURATION;
-const SECONDS_IN_YEAR = 60 * 60 * 24 * 365;
-const EPOCH_DURATION_IN_YEARS = EPOCH_DURATION / SECONDS_IN_YEAR;
-const BETA_PER_YEAR = 1 / 3;
-export const GENERATION_PER_SLOT = Math.pow(2, -17);
-const BOOTSTRAPPING_DURATION = 1154;
-const REWARDS_MANA_SHARE_COEFFICIENT = 2;
-export const EPOCH = 1154 + 1;
-
-// given t in seconds, returns the slot that time t belongs to
-function time_to_slot(t: number): number {
-  const slot = Math.floor(t / SLOT_DURATION);
-  return slot;
-}
-
-// given t in seconds, returns the epoch that time t belongs to
-function time_to_epoch(t: number): number {
-  const epoch = Math.floor(t / EPOCH_DURATION);
-  return epoch;
-}
 
 // given n in slots, returns the epoch that slot n belongs to
-function slot_to_epoch(n: number): number {
+function slotToEpoch(n: number): number {
   const epoch = Math.floor(n / SLOTS_IN_EPOCH);
   return epoch;
 }
 
 // returns the first slot of epoch n
-export function first_slot_of_epoch(n: number): number {
+export function getFirstSlotOfEpoch(n: number): number {
   const slot = n * SLOTS_IN_EPOCH;
   return slot;
 }
@@ -48,23 +24,27 @@ export function decay(value: number, n: number): number {
 }
 
 // returns the potential mana generated by holding value tokens from creationSlot to consumptionSlot
-export function potential_Mana(
+export function getPotentialMana(
   value: number,
   creationSlot: number,
   consumptionSlot: number,
   generationRate: number,
 ): number {
-  const i = slot_to_epoch(creationSlot);
-  const j = slot_to_epoch(consumptionSlot);
-  const n = j - i;
-  const slotAfterEpochi = first_slot_of_epoch(i + 1);
-  const firstSlotEpochJ = first_slot_of_epoch(j);
-  const d1 = slotAfterEpochi - creationSlot;
-  const d2 = consumptionSlot - firstSlotEpochJ;
+  const creationEpoch = slotToEpoch(creationSlot);
+  const consumptionEpoch = slotToEpoch(consumptionSlot);
+
+  const epochsLeftToConsumption = consumptionEpoch - creationEpoch;
+
+  const slotAfterCreationEpoch = getFirstSlotOfEpoch(creationEpoch + 1);
+  const firstSlotConsumptionEpoch = getFirstSlotOfEpoch(consumptionEpoch);
+
+  const d1 = slotAfterCreationEpoch - creationSlot;
+  const d2 = consumptionSlot - firstSlotConsumptionEpoch;
+  
   let potentialMana = 0;
-  if (n == 0) {
+  if (epochsLeftToConsumption == 0) {
     potentialMana = value * (consumptionSlot - creationSlot) * generationRate;
-  } else if (n == 1) {
+  } else if (epochsLeftToConsumption == 1) {
     potentialMana =
       value * d2 * generationRate + decay(value * d1 * generationRate, 1);
   } else {
@@ -72,8 +52,8 @@ export function potential_Mana(
       Math.exp(-BETA_PER_YEAR * EPOCH_DURATION_IN_YEARS) /
       (1 - Math.exp(-BETA_PER_YEAR * EPOCH_DURATION_IN_YEARS));
     const aux = value * generationRate * c * SLOTS_IN_EPOCH;
-    const potentialMana_n = decay(value * d1 * generationRate, n);
-    const potentialMana_n_1 = decay(aux, n - 1);
+    const potentialMana_n = decay(value * d1 * generationRate, epochsLeftToConsumption);
+    const potentialMana_n_1 = decay(aux, epochsLeftToConsumption - 1);
     const potentialMana_0 = value * d2 * generationRate + aux;
     potentialMana = potentialMana_n - potentialMana_n_1 + potentialMana_0;
   }
